requirements.txt
einops==0.3.0
fsspec==2023.9.0
ftfy==6.2.0
horovod==0.28.1
huggingface_hub==0.23.0
matplotlib==3.7.2
mmcv==2.0.1
mmengine==0.8.4
mmsegmentation==1.1.1
myutils==0.0.21
nltk==3.8.1
numpy==2.0.0
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.76
openpyxl==3.1.2
Pillow==10.4.0
pycocotools==2.0.7
regex==2023.8.8
safetensors==0.4.3
scipy==1.14.0
skimage==0.0
timm==0.4.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.2
transformers==4.37.2
